requirements.txt
# Flask and extensions
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Migrate>=4.0,<5.0
Flask-Login>=0.6,<1.0
Werkzeug>=3.0,<4.0

# Database
psycopg[binary]>=3.1,<4.0

# Image processing
Pillow>=10.4,<12.0

# WSGI server
gunicorn>=21.2,<23.0

# Utilities
python-dotenv>=1.0,<2.0

# Testing
pytest>=8.0
pytest-cov>=4.0
